+++ packages/ethereum/sdk/src/config/index.ts
@@ -21,6 +21,9 @@ import { rariMainnetConfig } from "./rari"
 import { zkatanaConfig } from "./zkatana"
 import { baseConfig } from "./base"
 import { baseSepoliaConfig } from "./base-sepolia"
+import { astarZKEVMConfig } from "./astar-zkevm"
+import { celoTestnetConfig } from "./testnet-celo"
+import { celoConfig } from "./celo"
 
 export const configDictionary: Record<EthereumNetwork, EthereumConfig> = {
 	mainnet: mainnetConfig,
@@ -42,8 +45,11 @@ export const configDictionary: Record<EthereumNetwork, EthereumConfig> = {
 	"testnet-rari": rariTestnetConfig,
 	rari: rariMainnetConfig,
 	zkatana: zkatanaConfig,
+	"astar-zkevm": astarZKEVMConfig,
 	base: baseConfig,
 	"base-sepolia": baseSepoliaConfig,
+	"testnet-celo": celoTestnetConfig,
+	celo: celoConfig,
 }
 
 export function getEthereumConfig(env: EthereumNetwork): EthereumConfig {