+++ packages/ethereum/web3-ethereum/src/domain.ts
@@ -8,5 +8,8 @@ export type Web3EthereumGasOptions = Partial<{
 export type Web3EthereumConfig = Web3EthereumGasOptions & {
 	web3: Web3
 	from?: string
-	alternateWeb3Instance?: Web3
+	/**
+   * Reserve nodes map in format { [chainId]: nodeUrl }
+   */
+	reserveNodes?: Record<number, string>
 }