+++ packages/transaction/src/ethereum/index.ts
@@ -76,6 +76,12 @@ Blockchain, TransactionResult
 				return `http://polygon-explorer.dev.rarible.int/transactions/${this.hash()}`
 			case "zkatana":
 				return `https://zkatana.blockscout.com/tx/${this.hash()}`
+			case "astar-zkevm":
+				return `https://astar-zkevm.explorer.startale.com/tx/${this.hash()}`
+			case "testnet-celo":
+				return `https://explorer.celo.org/alfajores/tx/${this.hash()}`
+			case "celo":
+				return `https://explorer.celo.org/mainnet/tx/${this.hash()}`
 			default:
 				throw new Error("Unsupported transaction network")
 		}